Add create account & library view functionality; update LibraryDao
CreateAccount.java -- creating account functionality created for librarian. Username has to be unique. If the new user registers when the librarian box is checked, they are brought to the "Librarian view". 

LibrarianVIew.java -- In this view, the librarian can add books, search/view books, and delete books. 

LibraryDao.java -- updated and added a few methods.